test(mock-db): modify drizzle db instance to use mock database in a test environment - updated db instance to use the testing db when process.env.VITEST is true - add support for 'TEST_DATABASE_PATH' & 'DATABASE_PATH' environmental variables - add '*.db', 'public/users' & '.env' to gitignore
So functions could operate natively in a test environment.

I tried the native mock driver of drizzle, but it gave me annoying errors, with only a paragraph about it in the docs.
So I tried the in-memory db of drizzle. Still, I couldn't find a native way to migrate the schema to the in-memory db without using the CLI, and migrating it manually isn't a choice
So the last choice is to use a separate database for testing for now.